Fix2

+++ M13_Practica2_MWC/src/main/java/com/mycompany/m13_practica2_mwc/Dispositivos.java
@@ -32,7 +32,7 @@ public class Dispositivos {
         return preuBase;
     }
 
-    public void setPreuBase(int preuBase) {
+    public void setPreuBase(double preuBase) {
         this.preuBase = preuBase;
     }
